requirements.txt
bambulabs_api==2.6.4
python-telegram-bot==22.5
telegramify-markdown==0.5.2